server/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-activatable-model==2.0.0
django-cors-headers==3.7.0
django-entity==5.0.1
django-manager-utils==2.0.1
django-query-builder==2.0.0
djangorestframework==3.12.2
fleming==0.6.0
idna==2.10
jsonfield==3.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python3-utils==0.4.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
wrapt==1.12.1
